+++ CHANGELOG.md
@@ -14,8 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Prevent context leaking through using context.applicationContext in constructors
+- Set current position for videoTimeStart and videoTimeEnd when customData event is sent
 - IVS Player: Ignore isLive property for IVS player since it is just a fallback in case the player
-  does not provide the information and IVS player doesn't differenciate between
+  does not provide the information and IVS player doesn't differentiate between
   live and source not loaded, so it cannot be used there. This streamlines the behaviour with other players.
 - ExoPlayer: More reliable tracking if player is muted